Extract and generalize collision code.
Fix error with resetting jump when sliding down wall.
Fix error with resetting vertical velocity when sliding down wall.
Error with being able to jump in the air when walking of edge is still prevalent.

+++ GitWorkflow/GitWorkflow/main.cpp
@@ -24,7 +24,7 @@ int main()
 	{
 		Time t2 = t1;
 		t1 = Clock::now();
-		const float dt = static_cast<float>((t1 - t2).count()) / 1e9;
+		const float dt = static_cast<float>((t1 - t2).count()) / static_cast<float>(1e9);
 
 		sf::Event event;
